Here are my two CMD cluster plots showing G vs G-Ks and G vs BP-RP for NGC2547, the Pleiades and the Hyades. I made a correction for the BP-RP flux excess and removed any targets that didn't satisfy equation 18 in Riello et al. (2021, A&A 649, A3).
I tried to make the program somewhat more functional by settting the
magnitudes and colours as generic variables rather than explicitly
defining their filters. This means at a later date we can more easily
compare, say, G versus G-J.

I'll just remind you that the method to get the polynomial fits for
the clusters is to fit a 2nd order polynomial through the whole CMD,
and then fit objects on the faint side of this line with a 4th order
polynomial. However, I noticed for the Pleiades that the isochrone
looked maybe 0.1-0.2 mag fainter than the observed sequence for the
late K-stars. To fix this I've move the 2nd order polynomial upwards
by 0.3 mag, and now the fit to the single stars seems to match the
observed set a little better.